Add lower type bounds to the entity argument X
Add lower type bounds to the entity argument X of the return value of
Path.get(PluralAttribute) and Path.get(MapAttribute). The absense of
these lower bounds prevent usage of the JPA criteria API for a
collection field f which is declared in an entity A which then is
extended by another entity B and used in the type bounds of TypedQuery.

closes #108

Signed-off-by: Karl-Philipp Richter <krichter@posteo.de>